+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Display warning only if pushing latest tag fails, as container image repository may have immutability enabled.
 - Taito CLI now supports PR environments when using Helm. That is, you can use `pr-NUMBER` as ENV to deploy pull-request version aside your dev version.
+- Postgres plugin now supports "db create" and "db drop" commands and database are create with the app specific database mgr user by default.
 
 ## 0.309.0